Move types to mui.d.ts file
